refactor : GNB 파트 리펙토링
